\name{mspec}
\alias{mspec.h}
\alias{mspec.c}
\docType{data}
\title{
Protein Mass Spectroscopy
}
\description{
For each blood serum sample, the intensity is recorded for a large number of time-of-flight values. \code{mspec.h} contains samples from 157 healthy patients, and \code{mspec.c} contains samples from 167 patients with cancer.
}
\usage{data("mass.spec")}
\format{
\code{mspec.h} is a matrix containing 157 rows and 2181 columns. \code{mspec.c} is a matrix containing 167 rows and 2181 columns.
}

\source{
Adam, B. L., Qu, Y., Davis, J. W., Ward, M. D., Clements, M. A., Cazares, L. H.,
Semmes, O. J., Schellhamm, P. F., Yasui, Y., Feng, Z. and Wright, G. L. W. Jr.
(2003) \emph{Cancer Research}. \bold{63}, 3609-3614.
}
\references{
Qiu, Y. and Chen, S. X. (2015) \emph{Journal of the American Statistical Association}. \bold{110.511}, 1160--1174.
}

